git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@7252 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Strategic/Map Screen Interface Map.cpp
@@ -4006,7 +4006,7 @@ void DisplayDistancesForHelicopter( void )
 	UINT32 uiCostRed = __max(0,gGameExternalOptions.usHelicopterBaseCostPerRedTile + gsSkyriderCostModifier);
 	uiTripCost = ( sNumSafeSectors * uiCostGreen ) + ( sNumUnSafeSectors * uiCostRed );
 
-	if( uiTripCost > LaptopSaveInfo.iCurrentBalance )
+	if( (INT32)uiTripCost > LaptopSaveInfo.iCurrentBalance )
 		SetFontForeground( FONT_LTRED );
 		
 	swprintf( sString, L"%d", uiTripCost );
